Issue #18: Refactored server code.
