+++ server/src/config/data.ts
@@ -1,5 +1,8 @@
 import { isDev } from './env';
 
+/** Default records count */
+export const defaultCount = 20;
+
 // TODO: Get the value from the environment
 export const recordsCount =
   process.env.recordsCount && !isNaN(Number(process.env.recordsCount))
