+++ server/src/lib/helpers/quotePgStr.ts
@@ -1,0 +1,4 @@
+/** Quotes a string for psql expression */
+export function quotePgStr(text: string) {
+  return text.replace(/'/g, "''");
+}
